Add control plane to SDK (#14)
This PR adds the Control Plane REST API calls for managing Indexes and
Collections.

Similar to the codegen for the Data Plane gRPC calls, this generates
code from the OpenAPI Spec for the Control Plane REST calls. These calls
are then wrapped in the SDK to reduce the surface area of the SDK to
support.