Try new strategy for better rendering of documentation with gopages (#45)
## Problem

Our current documentation is quite ugly (https://sdk.pinecone.io/go/).
We generated it by running `godoc -url`, which basically prints standard
html to a file. We then pushed this static file to `sdk-docs` via that
repo's README instructions.

## Solution

There is a Go module called `gopages`:
https://johnstarich.com/go/gopages/pkg/github.com/johnstarich/go/gopages/
that seemingly renders things better. We should try that to see if it
looks nicer.

## Type of Change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] This change requires a documentation update
- [x] Infrastructure change (CI configs, etc)
- [ ] Non-code change (docs, etc)
- [ ] None of the above: (explain here)

## Test Plan

Need to merge to main, then trigger the documentation workflow in the
`sdk-docs` repo to see the rendering.