Add `Namespace` to List, Query, and Fetch vector responses (#52)
## Problem
@jseldess noticed that the Go client doesn't return an explicit
`Namespace` with data plane operations. This would be helpful for the
end user as `Namespace` is configured when a user targets an index, so
having the `Namespace` associated with the specific operation would be
useful for validating the namespace where the data was retrieved.
Additionally, other SDKs such as TypeScript return namespace as a part
of the response object.

## Solution
- Add `Namespace` as a field in `FetchVectorsResponse`,
`ListVectorsResponse`, and `QueryVectorsResponse`.
- Update associated unit tests, mainly around marshaling of these
structs.

We do not use `omitempty` for `Namespace` as the `""` namespace has
meaning as the default.

## Type of Change
- [ ] Bug fix (non-breaking change which fixes an issue)
- [X] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] This change requires a documentation update
- [ ] Infrastructure change (CI configs, etc)
- [ ] Non-code change (docs, etc)
- [ ] None of the above: (explain here)

## Test Plan
Make sure CI unit / integration tests are passing. Make fetch, query, or
list operations and confirm the appropriate `Namespace` is returned
depending on which index is targeted.