Add `Host` normalization when building `IndexConnection` (#56)
## Problem
We had an issue that was logged around trying to call `Client.Index()`
with a `Host` value including an `http://` or `https://` schema. The
control plane does not attach these schemes, but when building an
`IndexConnection` a user needs to provide a host and they could make
this mistake. `grpc-go` is not expecting a `target` in `NewClient` that
includes a scheme.

We could do a bit of normalization to handle the `Host` more thoroughly.

## Solution
- Add `normalizeHost` helper function in `index_connection.go`. This
strips out `http://` & `https://` prefixes, and attaches the port `:443`
unless the user has provided one directly. We want to leave some
flexibility for making RPCs in specific scenarios.
- Add unit tests for `normalizeHost`.

## Type of Change
- [ ] Bug fix (non-breaking change which fixes an issue)
- [X] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] This change requires a documentation update
- [ ] Infrastructure change (CI configs, etc)
- [ ] Non-code change (docs, etc)
- [ ] None of the above: (explain here)

## Test Plan
CI unit & integration tests should pass as normal.

To test manually, you can try and prepend schemes to the Host when
calling `client.Index()`

```Go
indexConn, err := client.Index(pinecone.NewIndexConnParams{Host: fmt.Sprintf("https://%s", index.Host)})
if err != nil {
    panic(err)
}

stats, err := indexConn.DescribeIndexStats(context.Background())
```