Pass `currentVersion` in `release` workflow (#70)
## Problem
I tried running the new `release` action and it errored when trying to
push to `main` and when generating the new version tag due to 3 things:
- `main` branch doesn't allow pushes without PRs, we need to disable
that to allow more flexibility from CI (we do this in the python and
typescript repos).
- The `currentVersion` retrieved using `git describe` was not being
passed to the `bump-version` action, so we ended up with NaN in the
output.
- The `prereleaseSuffix` defaulting to "alpha1" and then being passed to
`bump-version` was an issue because it was being appended whether or not
`isPrerelease` was true.

Failed run:
https://github.com/pinecone-io/go-pinecone/actions/runs/10529783970/job/29178282163

## Solution
- Add explicit `SEMVER_VERSION` and `PRERELEASE_SUFFIX` variables to the
`release` workflow. Make sure these values are set in the
`get_tag_version` and `set_prerelease_suffix` steps.
- Pass the `currentVersion` value to `bump-version` properly. 

## Type of Change
- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] This change requires a documentation update
- [X] Infrastructure change (CI configs, etc)
- [ ] Non-code change (docs, etc)
- [ ] None of the above: (explain here)

## Test Plan
Push changes up and rerun the version bump.