Refactor `/codegen/build-clients.sh` to support new `apis` structure, regenerate for `2024-10` (#76)
## Problem
The upstream `/apis` submodule has undergone some changes since we last
regenerated the client core. Our `/codegen/build-clients.sh` script no
longer handles the new module naming or structure properly.

Additionally, we want to regenerate off the upcoming API version so we
can pull in changes for things like Rerank.

This PR is **merging to a release candidate (RC) branch:
`release-candidate/2024-10`**. We will track upcoming major version
changes in this branch and use it for releasing dev builds.

## Solution
- Refactor `codegen/build-clients.sh` script to support the newer
structure for the upcoming API version, specifically `inference` is its
own module, and `db_control`/`db_data` as separate modules.
- Update imports and exports for `control` -> `db_control`, and
`control` -> `inference` for the `Embed()` method on `InferenceService`.

We made the decision to have `InferenceService` as its own module inside
`Client` because of the split in the API spec, so this was fairly easy
to refactor around here.

## Type of Change
- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] This change requires a documentation update
- [X] Infrastructure change (CI configs, etc)
- [ ] Non-code change (docs, etc)
- [ ] None of the above: (explain here)

## Test Plan
The code changes under `/internal/gen/*` are all from running the new
`/codegen/build-clients.sh` script. I changed files in the submodule for
`apis` manually for generating, but it aligns with this PR:
https://github.com/pinecone-io/apis/pull/138

Make sure CI build + tests pass for the new core stuff here.