Implement Bulk Import, Regenerate core for `2024-10` API (#79)
## Problem
We are releasing a new version of the API this month: `2024-10`.

There are 3 primary new features that are included in this release:
- Import
- Inference
  - Embed
  - Rerank

This PR implements the operations to support import. Sorry about the
size, but you can basically ignore all of the generated code under
`internal/gen` unless you're curious about the new structure of the
generated core files. Follow the `codgen/build-clients.sh` script for
those details.

## Solution
Since the import operations are technically part of the data plane but
only supported via REST, they are represented in the OpenAPI spec and
not our protos file. Because of this, we need to change a few things to
support these operations `Client` and `IndexConnection` structs to
support these operations because traditionally the code
`IndexConnection` wraps was targeting gRPC-only db data operations. We
now need to generate rest code for the data plane as well so we can
interact with imports.

- Update the `codegen/build-clients.sh` script to handle building new
modules for both `internal/gen/db_data/grpc` and
`internal/gen/db_data/rest`.
- Update `Client` struct and move `NewClientBaseParams` into a field
that can be shared more easily when constructing the `IndexConnection`.
- Add `buildDataClientBaseOptions` to handle constructing the necessary
rest client options for the underlying `dbDataClient`.
- Add an `ensureHostHasHttps` helper as we need to make sure this is
present for the index `Host` that's passed, which was not necessary for
grpc.
- Update `Index` method to handle calling `buildDataClientBaseOptions`
and passes the new client into `newIndexConnection`.
- Update `IndexConnection` to support both REST and gRPC interfaces
under the hood (`restClient`, `grpcClient`).
- Update `newIndexConnection` to support attaching the new `restClient`
to the `IndexConnection` struct.
- Update `IndexConnection` to support all import operations:
`StartImport`, `ListImports`, `DescribeImport`, `CancelImport`.
- Add end-to-end integration test for validating the import flow against
serverless indexes.
- Some nitpicky code cleanup, renaming of things around the new rest vs.
grpc paradigm, etc.

## Type of Change
- [ ] Bug fix (non-breaking change which fixes an issue)
- [X] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] This change requires a documentation update
- [ ] Infrastructure change (CI configs, etc)
- [ ] Non-code change (docs, etc)
- [ ] None of the above: (explain here)

## Test Plan
`just test` - make sure CI passes

To see examples of how to use the new methods, check the doc comments.


---
- To see the specific tasks where the Asana app for GitHub is being
used, see below:
  - https://app.asana.com/0/0/1208325183834377
  - https://app.asana.com/0/0/1208541827330963