Add `EmbedResponse` type, regenerate core for `2024-10` (#83)
## Problem
There have been a few small PRs that have gone into the `apis` repo that
we need to regenerate code for:

- https://github.com/pinecone-io/apis/pull/175
- https://github.com/pinecone-io/apis/pull/171

We're also returning the generated `EmbeddingsList` type for the `Embed`
method rather than a custom type like we're doing elsewhere.

## Solution
- Add new `EmbedResponse`, and `Embedding` types and return from
`Embed`. Add doc comments.
- Regenerate core SDK code from `2024-10`

## Type of Change
- [ ] Bug fix (non-breaking change which fixes an issue)
- [X] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] This change requires a documentation update
- [ ] Infrastructure change (CI configs, etc)
- [ ] Non-code change (docs, etc)
- [ ] None of the above: (explain here)

## Test Plan
Make sure CI is successful.