Regenerate core for `2025-01` (#92)
## Problem
In preparation for the upcoming API version release, we need to
regenerate the client code off of the specifications for `2025-01`.

## Solution
This PR regenerates the client core from `main`, and merges the updates
into the RC branch: `2025-01`. This branch is currently synced with
`main`.

## Type of Change
- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [X] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] This change requires a documentation update
- [ ] Infrastructure change (CI configs, etc)
- [ ] Non-code change (docs, etc)
- [ ] None of the above: (explain here)

## Test Plan
Ran unit and integration tests locally to validate changes. Make sure CI
is green before merging into the RC branch.


---
- To see the specific tasks where the Asana app for GitHub is being
used, see below:
  - https://app.asana.com/0/0/1208897809680471