Update `go.mod` for v3, review feedback (#97)
## Problem
In order to release `v3.0.0` the `go.mod` file and associated imports
need to be updated. There was also some feedback from a previous PR that
I wanted to implement.

## Solution
- Update `go.mod` to bump `module` to v3 for the upcoming release.
Update relevant imports.
- Make `Metric` and `DeletionProtection` pointers to signify optionality
in `CreateServerlessIndexRequest` and `CreatePodIndexRequest`. Update
tests, README, etc.

## Type of Change
- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [X] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] This change requires a documentation update
- [ ] Infrastructure change (CI configs, etc)
- [ ] Non-code change (docs, etc)
- [ ] None of the above: (explain here)

## Test Plan
CI unit + integration