Bump dependency: `golang.org/x/net@v0.35.0` (#111)
Dependabot opened a PR for bumping the `golang.org/x/net` dependency
from 0.33.0 -> 0.38.0. However, in order to make that jump, we'd need to
bump the Go version our module supports from go 1.21 -> go 1.24. For
consumers, this would be a breaking change and they'd need to upgrade
their version of Go to use the SDK.

At the moment, I don't think is necessary for this release. For now I've
updated to `golang.org/x/net v0.35.0` which is the last version which
supports < `go v1.23`:
- https://cs.opensource.google/go/x/net/+/refs/tags/v0.35.0:go.mod
- https://cs.opensource.google/go/x/net/+/refs/tags/v0.36.0:go.mod

Bump `golang.org/x/net` -> `v0.35.0`.

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] This change requires a documentation update
- [X] Infrastructure change (CI configs, etc)
- [ ] Non-code change (docs, etc)
- [ ] None of the above: (explain here)

CI