Update `README` to use v4 instead of v3 (#113)
## Problem
The major release for v4.0.0 updated the package route to 

## Solution
Update code examples in the `README` to use `/v4/` instead of `/v3/`.
There was also a spot in the justfile where we spin up a local server
for serving docs that needed a tweak too.

## Type of Change
- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] This change requires a documentation update
- [ ] Infrastructure change (CI configs, etc)
- [X] Non-code change (docs, etc)
- [ ] None of the above: (explain here)

## Test Plan
N/A