fix ensure https function (#114)
## Problem

Describe the purpose of this change. What problem is being solved and
why?
https double concatenated on go client search calls. the has prefix
function input is string followed by the prefix not the other way
around.

## Solution

fix the bug on has prefix 
## Type of Change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] This change requires a documentation update
- [ ] Infrastructure change (CI configs, etc)
- [ ] Non-code change (docs, etc)
- [ ] None of the above: (explain here)

## Test Plan

Describe specific steps for validating this change.

Co-authored-by: Austin DeNoble <austin.d@pinecone.io>