Add `Host` as an optional override parameter for `AdminClient` (#123)
## Problem
For `Client` we allow users to pass a custom `Host` parameter along with
`NewClientParams`. This is useful for development work, specifically for
targeting staging environments through the SDK.

`AdminClient` uses the same API as `Client`, but currently does not
support passing a `Host`.

## Solution
In the same vein as `Client`, add `Host` to `NewAdminClientParams`, and
implement the same logic we use in `Client`.

## Type of Change
- [ ] Bug fix (non-breaking change which fixes an issue)
- [X] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] This change requires a documentation update
- [ ] Infrastructure change (CI configs, etc)
- [ ] Non-code change (docs, etc)
- [ ] None of the above: (explain here)

## Test Plan
CI - unit and integration tests