Fix `ConfigureIndex` when passing `ReadCapacity` (#129)
## Problem
There was a small bug in the OpenAPI specification for `configure_index`
requests that patch the index's `read_capacity` configuration. You can
see the upstream fix for the API schemas here:
https://github.com/pinecone-io/apis/pull/411

The gist is `node_type`, `shards`, and `replicas` were all marked as
required in schemas that used for both create and update operations.
What this meant was the generated Go code was inflexibly for update
operations. Because the fields were marked as required, the zero-value
for `shards` or `replicas` would end up being sent for patch requests if
the user didn't provide them.

Additionally, the business logic around patching `read_capacity` was not
taking into account the previous configuration on update, which is
useful for validating the patch request.

## Solution
- Add new `ReadCapacityOnDemandConfig`. This is an empty struct used in
the `ReadCapacityParams` union type. Originally, I left this out as
leaving `ReadCapacityParams` off entirely will default the index to
OnDemand. However, we also need a way to represent configuring an index
to convert from Dedicated -> OnDemand, and the absence of a Dedicated
`read_capacity` isn't enough to signal the swap to OnDemand. This also
better aligns with how the union types are represented in the spec.
- Update `ReadCapacityDedicated` and `ReadCapacityManualScaling` to
account for optionality (pointers) for `NodeType`, `Replicas`, and
`Shards`.
- Add new `patchReadCapacity` function for validating the patch request.
- Update `readCapacityParamsToReadCapacity` to better handle error
scenarios.
- Update unit tests and integration tests to account for changes to
types.

NOTE: This change will cause changes to how users interface with the DRN
operations - they will need to pass pointers for `NodeType`, `Shards`,
and `Replicas`. Additionally, they will need to deal with pointers for
these values in responses containing `ReadCapacity`.

## Type of Change
- [X] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] This change requires a documentation update
- [ ] Infrastructure change (CI configs, etc)
- [ ] Non-code change (docs, etc)
- [ ] None of the above: (explain here)

## Test Plan
CI - unit & integration tests