Update `Client.ConfigureIndex` documentation (#130)
## Problem
The structs used in `ReadCapacityParams` for `Client.ConfigureIndex` may
be a little confusing. It would be helpful to add some clarity about how
they're used, and which fields are required.

## Solution
Add details around `OnDemand` vs. `Dedicated`, and which fields are
required when configuring a newly-dedicated index.

## Type of Change
- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] This change requires a documentation update
- [ ] Infrastructure change (CI configs, etc)
- [X] Non-code change (docs, etc)
- [ ] None of the above: (explain here)

## Test Plan
CI - integration & unit tests