Return all fields for `NamespaceDescription` on `CreateNamespace` call (#132)
## Problem
Noticed that the `NamespaceDescription.Schema` field was not being
properly returned from the `CreateNamespace` method on `IndexConnection`

## Solution
Return `Schema` and `IndexedFields` for `NamespaceDescription` values
returned from the `CreateNamespace` method. This is the same translation
between generated grpc code and the SDK wrapper as in
`DescribeNamespace`.

## Type of Change
- [X] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] This change requires a documentation update
- [ ] Infrastructure change (CI configs, etc)
- [ ] Non-code change (docs, etc)
- [ ] None of the above: (explain here)

## Test Plan
CI - unit & integration tests. This change is additive, and shouldn't
impact existing functionality in any way.